Cockpit-backend: Refactor into multiple files
This makes the code a lot more readable.
